refactor python to move common class/regex into utils
