+++ eng/versioning/.gitignore
@@ -1,0 +1,6 @@
+# Python cache
+__pycache__/
+*.pyc
+.pytest_cache
+.mypy_cache
+.cache
